Catching Throwable instead of Exception in the Runnable passed to the scheduleAtFixedRate method invoked in the Scheduler.
According to the https://docs.oracle.com/javase/8/docs/api/java/util/concurrent/ScheduledExecutorService.html , "If any execution of the task encounters an exception, subsequent executions are suppressed."
In case of an Error (as opposed to Exception) the scheduler used to stop working, as Error does not extend Exception, but Throwable. Now it will carry on.